betty style is verified for task 1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
 void print_binary(unsigned long int n)
 {
 	int bits = sizeof(unsigned long int) * 8;
-	unsigned long int mask = (unsigned long int)1 << (bits - 1);
+	unsigned long int mask = 1 << (bits - 1);
 	int leadingZeros = 1;
 
 	for (int i = 0; i < bits; i++)
